Completed the implementation of the methods callbacks on the event system - Now the methods can be inserted on the EventDispatcher - Fixed the RemoveEventListener and HasEventListener functions - Fixed some minor bugs and issues

+++ Shard2D/Shard2D/src/Framework/Events/EventsManager.cpp
@@ -25,7 +25,7 @@ namespace Shard2D
 		int _actualMouseX = 0;
 		int _actualMouseY = 0;
 		SDL_GetMouseState(&_actualMouseX, &_actualMouseY);
-		mousePosition = Vector2(_actualMouseX, _actualMouseY);
+		mousePosition = Vector2((float)_actualMouseX, (float)_actualMouseY);
 
 		// Setup the arrays
 		eventsToDispatch.clear();
